Rearchitectured System. Line Tool Working

+++ Core/API/Vertex2.h
@@ -3,12 +3,12 @@
 
 class Vertex2 {
 public:
-	double x;
-	double y;
+	float x;
+	float y;
 
 	inline Vertex2() : x(0.0), y(0.0) {}
-	inline Vertex2(double _x, double _y) : x(_x), y(_y) {}
-	inline Vertex2(std::initializer_list<double> list) : x(*list.begin()), y(*(list.begin() + 1)) {}
+	inline Vertex2(float _x, float _y) : x(_x), y(_y) {}
+	inline Vertex2(std::initializer_list<float> list) : x(*list.begin()), y(*(list.begin() + 1)) {}
 
 	inline bool operator==(const Vertex2& rhs) 
 	{
